requirements.txt
requests==2.25.1
pyyaml==6.0.1